Update logic of normal mode entry combinations
jj and kv were being triggered even when typing "juj" or "kev" if fast
enough. Now, any key in between will prevent the trigger, but there
is no timeout making it impossible to type jj or kv characters without
something inbetween. This was considered a necessary tradeoff,
considering the rarity of these character combinations in english.